separated modal form into 2 separate components
